+++ Realm.Shared/Schema/ObjectSchema.cs
@@ -27,6 +27,9 @@ using System.Reflection;
 
 namespace Realms.Schema
 {
+    /// <summary>
+    /// Public description of a class stored in a Realm, as a collection of managed Property objects.
+    /// </summary>
     [DebuggerDisplay("Name = {Name}, Properties = {Count}")]
     public class ObjectSchema : IReadOnlyCollection<Property>
     {
@@ -82,7 +85,7 @@ namespace Realms.Schema
                 var schemaProperty = new Property
                 {
                     Name = property.GetCustomAttribute<MapToAttribute>()?.Mapping ?? property.Name,
-                    IsObjectId = property.GetCustomAttribute<ObjectIdAttribute>() != null,
+                    IsPrimaryKey = property.GetCustomAttribute<PrimaryKeyAttribute>() != null,
                     IsIndexed = property.GetCustomAttribute<IndexedAttribute>() != null,
                     PropertyInfo = property
                 };