+++ Tests/IntegrationTests.Shared/PrimaryKeyAttribute.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace IntegrationTests.Shared
+{
+    class PrimaryKeyAttribute : Attribute
+    {
+    }
+}