+++ wrappers/src/realm_error_type.hpp
@@ -67,6 +67,8 @@ namespace realm {
 
         RealmRowDetached = 21,
 
+        RealmTableHasNoPrimaryKey = 22,
+
         StdArgumentOutOfRange = 100,
 
         StdIndexOutOfRange = 101,